--- a/src/main.rs
+++ b/src/main.rs
@@ -1,11 +1,11 @@
 // pub mod bitpack;
 // pub mod memory;
 // pub mod machine;
-use std::env;
 use rum::machine;
+use std::env;
 
 fn main() {
-	let args: Vec<String> = env::args().collect();
+    let args: Vec<String> = env::args().collect();
     let filename = &args[1];
     let instructions = machine::boot(filename);
     machine::run(instructions);