--- a/src/memory.rs
+++ b/src/memory.rs
@@ -1,51 +1,44 @@
-// pub mod memory {
-use std::mem;
+const PROGRAM_ADDRESS: usize = 0;
+
 #[derive(Debug)]
 pub struct Memory {
     pool: Vec<usize>,
     heap: Vec<Vec<u32>>,
 }
 
-const PROGRAM_ADDRESS: usize = 0;
 impl Memory {
     // create a new Memory, comprising a pool of reusable IDs
     // and a heap of UM words, populated with the instructions
     // as segment 0
     pub fn new(instructions: Vec<u32>) -> Memory {
-        Memory { pool: Vec::new(), heap: vec![instructions] }
+        Memory { pool: vec![], heap: vec![instructions] }
     }
 
     // allocate and initalize (as all 0s) a memory segment.
     // returns the segment ID
     pub fn allocate(&mut self, size: usize) -> usize {
-        let space = vec![0_u32; size];
+        let space = vec![0; size];
         // can we reuse a previously unmapped segment id?
         if self.pool.len() == 0 {
             self.heap.push(space);
             self.heap.len() - 1
         } else {
             let address = self.pool.pop().expect("No segment ID available");
-
-            mem::replace(
-                self.heap
-                    .get_mut(address)
-                    .expect("Memory was not previously allocated"),
-                space,
-            );
-
+            assert!(address < self.heap.len(), "invalid address in pool");
+            self.heap[address] = space;
             address
         }
     }
 
     // deallocate the memory at the given address.
     pub fn deallocate(&mut self, address: usize) {
-        self.pool.push(address);
-        mem::replace(
-            self.heap
-                .get_mut(address)
-                .expect("Memory was not previously allocated"),
-            Vec::new(),
+        assert!(
+            address < self.heap.len(),
+            "invalid address {}, cannot deallocate",
+            address,
         );
+        self.pool.push(address);
+        self.heap[address] = vec![];
     }
 
     // supply contents of the memory at the given address if
@@ -57,6 +50,7 @@ impl Memory {
             None => panic!("Segment unmapped!"),
         }
     }
+
     // get the instruction word corresponding to the given program counter
     // This may have high overhead...
     #[inline]
@@ -72,13 +66,13 @@ impl Memory {
     pub fn store(&mut self, seg_id: usize, address: usize, value: u32) {
         let memory =
             self.heap.get_mut(seg_id).expect("Memory was unallocated");
-
-        mem::replace(
-            memory
-                .get_mut(address)
-                .expect("No value present at given address"),
-            value,
+        assert!(
+            address < memory.len(),
+            "invalid address {} for segment {}",
+            address,
+            seg_id
         );
+        memory[address] = value;
     }
 
     // replace the program with the vector at the given address
@@ -88,13 +82,8 @@ impl Memory {
             .get(seg_id)
             .expect("Found no program at the given address")
             .clone();
-
-        mem::replace(
-            self.heap
-                .get_mut(PROGRAM_ADDRESS)
-                .expect("Found no existing program"),
-            program,
-        );
+        // Never panics, PROGRAM_ADDRESS is to be present by construction.
+        // (And because `heap` never shrinks.)
+        self.heap[PROGRAM_ADDRESS] = program;
     }
 }
-// }