--- a/umbin/expected/midmark
+++ b/umbin/expected/midmark
@@ -1,0 +1,8 @@
+read 120440 bytes from umbin/midmark.um
+ == UM beginning stress test / benchmark.. ==
+4.   12345678.09abcdef
+3.   6d58165c.2948d58d
+2.   0f63b9ed.1d9c4076
+1.   8dba0fc0.64af8685
+0.   583e02ae.490775c0
+Benchmark complete.