master:

```
[andrew@krusty rum]$ hyperfine 'rum-master umbin/midmark.um'
Benchmark #1: rum-master umbin/midmark.um
  Time (mean ± σ):     234.5 ms ±   4.2 ms    [User: 232.4 ms, System: 1.9 ms]
  Range (min … max):   227.5 ms … 242.3 ms    13 runs
```

This PR:

```
[andrew@krusty rum]$ hyperfine 'rum umbin/midmark.um'
Benchmark #1: rum umbin/midmark.um
  Time (mean ± σ):     196.7 ms ±   7.7 ms    [User: 194.2 ms, System: 2.3 ms]
  Range (min … max):   191.2 ms … 216.5 ms    15 runs
```

Which is good for about a 16% performance improvement.

As far as I can tell, the bottleneck at this point is the double pointer indirection in the store/load methods, which are by far the most frequent instruction (except for `LoadValue`). 

I would definitely like to see the C version that is 50% faster. :-) Does it use the same allocation strategy? Otherwise, I have a suspicion that the C code is doing a lot more unchecked index access than the Rust version, which would result in a machine that exhibits undefined behavior based on its input. That is, one could write a UM program that results in out-of-bounds access. This Rust program should not suffer from that flaw. With that said, a [careful reading of the spec](http://www.boundvariable.org/um-spec.txt) seems to suggest that UB is actually OK, since all it says (at the bottom) is that the machine "may" fail given some nonsensical program behavior.

I think the next step I would take here would be to eliminate the double indirection, although it would take a bit of work to do that in safe code while enabling the reuse of allocations. (You're essentially writing an allocator at that point.)