@ndaniels 2020-05-07
The refactoring is really helpful to my learning. Thanks!